+++ DemoApp/StreamChat/StreamChatWrapper+DemoApp.swift
@@ -57,6 +57,9 @@ extension StreamChatWrapper {
         Components.default.messageActionsVC = DemoChatMessageActionsVC.self
         Components.default.reactionsSorting = { $0.type.position < $1.type.position }
         Components.default.messageLayoutOptionsResolver = DemoChatMessageLayoutOptionsResolver()
-        Components.default.mixedAttachmentInjector.register(.location, with: LocationAttachmentViewInjector.self)
+        
+        if AppConfig.shared.demoAppConfig.isLocationAttachmentsEnabled {
+            Components.default.mixedAttachmentInjector.register(.location, with: LocationAttachmentViewInjector.self)
+        }
     }
 }